- Refactor script to improve readability and maintainability. - Add comments to explain the purpose of each section. - Check for user 'icaoberg' before executing Conda environment setup. - Move processed JSON file to specific directory only if it exists. - Handle the case where the user running the script is not 'icaoberg'.

+++ update-summary.sh
@@ -3,21 +3,31 @@
 #SBATCH -n 25
 #SBATCH --mem=50000M
 
-# this is only mean to aid @icaoberg run this as a cronjob
+# Description:
+# This script is designed to aid @icaoberg in running it as a cronjob.
 
-. /bil/users/icaoberg/miniconda3/etc/profile.d/conda.sh
-conda activate base
-python ./update-summary.py
-python ./create_list_file.py
+# Load Conda environment
+if [ "$USER" = "icaoberg" ] && [ -f "/bil/users/icaoberg/miniconda3/etc/profile.d/conda.sh" ]; then
+	. /bil/users/icaoberg/miniconda3/etc/profile.d/conda.sh
+	conda activate base
 
-sed -i 's|"Allen Institute for Brain Science "|"Allen Institute for Brain Science"|g' today.json
-sed -i 's|"award_number":null|"award_number":"Unavailable"|g' today.json
-sed -i 's|"Mouse"|"mouse"|g' today.json
-sed -i 's|"Human"|"human"|g' today.json
-sed -i 's|"Marmoset"|"marmoset"|g' today.json
-sed -i 's|"other"|"Other"|g' today.json
+	# Execute Python scripts
+	python ./update-summary.py
+	python ./create_list_file.py
 
-if [ -f /bil/data/inventory/daily/reports/today.json ]; then
-	rm -f /bil/data/inventory/daily/reports/today.json
-	cp today.json /bil/data/inventory/daily/reports/
+	# Modify JSON file ('today.json') using 'sed' commands
+	sed -i 's|"Allen Institute for Brain Science "|"Allen Institute for Brain Science"|g' today.json
+	sed -i 's|"award_number":null|"award_number":"Unavailable"|g' today.json
+	sed -i 's|"Mouse"|"mouse"|g' today.json
+	sed -i 's|"Human"|"human"|g' today.json
+	sed -i 's|"Marmoset"|"marmoset"|g' today.json
+	sed -i 's|"other"|"Other"|g' today.json
+
+	# Move processed JSON file to a specific directory if it exists
+	if [ -f /bil/data/inventory/daily/reports/today.json ]; then
+		rm -f /bil/data/inventory/daily/reports/today.json
+		cp today.json /bil/data/inventory/daily/reports/
+	fi
+else
+	echo "Only user icaoberg can run this script"
 fi
